Fix mount options for vfat filesystems on automounting
Hot-plug device monitoring is done by udev. On OpenPandora, the file
/etc/udev/rules.d/local.rules guide udev to call the script
/etc/udev/scripts/mount.sh to handle aumounting. This is the script
changed in this commit.

The change makes the use of blkid (already used by the script) to
probe the device type prior to mounting. If a device type of vfat is
detected, then the utf8 argument is append to the list of mount
options. This option is important to provide proper Unicode support
for filenames. More about this argument can be found at the
following link:

https://www.kernel.org/doc/Documentation/filesystems/vfat.txt

The several tests were done using a real OpenPandora hardware.

One strange thing to notice is the blkid variable at the
beginning of the file. On the repo, it is blkid="/sbin/blkid", but on
the OpenPandora, it is blkid="/usr/sbin/blkid".

I did no changes to this line.